pyproject.toml
[project]
name = "smart-ragflow"
version = "0.1.0"
description = "智能问答系统框架"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "langchain",
    "langgraph",
    "pydantic",
    "pydantic-settings",
    "opensearch-py",
    "aiohttp",
    "redis",
    "minio",
    "python-dotenv",
    "httpx",
    "numpy",
    "tiktoken",
    "sqlalchemy>=2.0.0",
    "pymysql>=1.1.0",
    "python-multipart",
    "pdfplumber",
    "python-docx",
    "openpyxl",
    "python-pptx",
]

[project.scripts]
parser = "backend_parser.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend_QA", "backend_parser", "backend_common", "backend_admin", "frontend_QA", "frontend_admin"]
